+++ jvm/src/main/scala/com/nawforce/apexlink/cst/TypeReference.scala
@@ -17,6 +17,7 @@ package com.nawforce.apexlink.cst
 import com.nawforce.apexlink.names.TypeNames
 import com.nawforce.apexlink.names.TypeNames._
 import com.nawforce.pkgforce.names.{EncodedName, Name, Names, TypeName}
+import com.nawforce.pkgforce.path.PathLocation
 import com.nawforce.runtime.parsers.CodeParser
 import io.github.apexdevtools.apexparser.ApexParser.{
   TypeArgumentsContext,
@@ -26,6 +27,7 @@ import io.github.apexdevtools.apexparser.ApexParser.{
 }
 
 import scala.collection.immutable.ArraySeq
+import scala.collection.mutable
 
 trait CSTTypeName {
   def typeArguments(): CSTTypeArguments
@@ -33,6 +35,11 @@ trait CSTTypeName {
   def isSet: Boolean
   def isMap: Boolean
   def getIdText: Option[String]
+
+  /** Source location of the identifier of this segment, if the adapter can provide one. Collection
+    * keywords (List/Set/Map) and synthetic types have no identifier so return None.
+    */
+  def idLocation: Option[PathLocation] = None
 }
 
 trait CSTTypeReference {
@@ -69,6 +76,8 @@ private[cst] object ANTLRCST {
     override def isMap: Boolean  = Option(typeName.MAP()).nonEmpty
     override def getIdText: Option[String] =
       Option(typeName.id()).map(id => Option(id).map(_.getText).getOrElse(""))
+    override def idLocation: Option[PathLocation] =
+      Option(typeName.id()).flatMap(id => CST.sourceContext.value.map(_.getLocation(id)))
   }
 
   private[cst] class ANTLRTypeReference(typeRef: TypeRefContext) extends CSTTypeReference {
@@ -81,6 +90,11 @@ private[cst] object ANTLRCST {
 
 object TypeReference {
 
+  /** Accumulator for the located type occurrences of a written type reference. A null accumulator
+    * disables collection so the common construction path does no extra work.
+    */
+  private type Occurrences = mutable.ArrayBuffer[SourceTypeOccurrence]
+
   def construct(typeRefs: List[TypeRefContext]): List[TypeName] = {
     typeRefs.map(x => TypeReference.construct(x))
   }
@@ -90,6 +104,27 @@ object TypeReference {
   }
 
   def construct(typeRefOpt: Option[CSTTypeReference]): TypeName = {
+    build(typeRefOpt, null)
+  }
+
+  /** Construct a type name along with the located occurrences of each explicitly written component
+    * of the reference, for use in source level accessibility validation.
+    */
+  def constructWithOccurrences(
+    typeRef: TypeRefContext
+  ): (TypeName, ArraySeq[SourceTypeOccurrence]) = {
+    constructWithOccurrences(Option(typeRef).map(new ANTLRCST.ANTLRTypeReference(_)))
+  }
+
+  def constructWithOccurrences(
+    typeRefOpt: Option[CSTTypeReference]
+  ): (TypeName, ArraySeq[SourceTypeOccurrence]) = {
+    val accum    = new mutable.ArrayBuffer[SourceTypeOccurrence]()
+    val typeName = build(typeRefOpt, accum)
+    (typeName, if (accum.isEmpty) SourceTypeOccurrence.empty else ArraySeq.from(accum))
+  }
+
+  private def build(typeRefOpt: Option[CSTTypeReference], accum: Occurrences): TypeName = {
     typeRefOpt
       .map { typeRef =>
         {
@@ -97,7 +132,15 @@ object TypeReference {
           val names: ArraySeq[CSTTypeName] = typeRef.typeNames()
 
           // Only decode head as rest can't legally be in EncodedName format
-          createTypeName(decodeName(names.head), names.tail).withArraySubscripts(arraySubs)
+          val typeName = createTypeName(decodeName(names.head, accum), names.tail, accum)
+          if (accum != null) {
+            // Record against the last written identifier, e.g. the 'Hidden' of 'Outer.Hidden'. The
+            // array subscripts are excluded so we validate the component, not the List wrapper.
+            names.lastOption
+              .flatMap(_.idLocation)
+              .foreach(location => accum.append(SourceTypeOccurrence(typeName, location)))
+          }
+          typeName.withArraySubscripts(arraySubs)
         }
       }
       .getOrElse(TypeNames.Void)
@@ -110,8 +153,8 @@ object TypeReference {
     else name.getIdText.map(Names(_)).getOrElse(Names.Empty)
   }
 
-  private def decodeName(name: CSTTypeName): TypeName = {
-    val params   = createTypeParams(name.typeArguments())
+  private def decodeName(name: CSTTypeName, accum: Occurrences): TypeName = {
+    val params   = createTypeParams(name.typeArguments(), accum)
     val typeName = getName(name)
     val encType  = EncodedName(typeName)
     if (encType.ext.nonEmpty)
@@ -121,22 +164,30 @@ object TypeReference {
   }
 
   @scala.annotation.tailrec
-  private def createTypeName(outer: TypeName, names: Seq[CSTTypeName]): TypeName = {
+  private def createTypeName(
+    outer: TypeName,
+    names: Seq[CSTTypeName],
+    accum: Occurrences
+  ): TypeName = {
     names match {
       case Nil => outer
       case hd +: tl =>
         createTypeName(
-          TypeName(getName(hd), createTypeParams(hd.typeArguments()), Some(outer)).intern,
-          tl
+          TypeName(getName(hd), createTypeParams(hd.typeArguments(), accum), Some(outer)).intern,
+          tl,
+          accum
         )
     }
   }
 
-  private def createTypeParams(typeArguments: CSTTypeArguments): Seq[TypeName] = {
+  private def createTypeParams(
+    typeArguments: CSTTypeArguments,
+    accum: Occurrences
+  ): Seq[TypeName] = {
     if (typeArguments.typeRefs().isEmpty)
       TypeName.emptySeq
     else
-      typeArguments.typeRefs().map(param => TypeReference.construct(Option(param)))
+      typeArguments.typeRefs().map(param => build(Option(param), accum))
   }
 }
 
@@ -144,5 +195,12 @@ object TypeList {
   def construct(typeList: TypeListContext): ArraySeq[TypeName] = {
     val types = CodeParser.toScala(typeList.typeRef())
     types.map(t => TypeReference.construct(t))
+  }
+
+  def constructWithOccurrences(
+    typeList: TypeListContext
+  ): (ArraySeq[TypeName], ArraySeq[SourceTypeOccurrence]) = {
+    val results = CodeParser.toScala(typeList.typeRef()).map(TypeReference.constructWithOccurrences)
+    (results.map(_._1), results.flatMap(_._2))
   }
 }